requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-decouple==3.4